Update README.md
In this project, we will implement a sorting program using a heap. Specifically, we will write a program that reads a list of commands from an input file and execute them, while maintaining a min heap and a max heap. Duplicate data can be inserted into the heap.

There are three commands that need to be supported by the program:

(1) insert

This command inserts an integer into both the min heap and the max heap. If the line reads “INSERT 3”, the program should insert a new element with data = 3 to the min heap and the max heap.

(2) print data in the ascending order

The command “ASCEND” prints the elements in the heap in the ascending order. At this command, the program should write a line which contains the sorted elements in the output file named “hw2_result.txt”. (Try running the given binary file.)